+++ torchbearer/metrics/default.py
@@ -7,6 +7,7 @@ following strings:
 - '`binary_acc`': The :class:`.BinaryAccuracy` metric
 - '`cat_acc`': The :class:`.CategoricalAccuracy` metric
 - '`top_5_acc`': The :class:`.TopKCategoricalAccuracy` metric
+- '`top_10_acc`': The :class:`.TopKCategoricalAccuracy` metric with k=10
 - '`mse`': The :class:`.MeanSquaredError` metric
 - '`loss`': The :class:`.Loss` metric
 - '`epoch`': The :class:`.Epoch` metric